Rename setName to setTaskName
I could refactor the validateProject and validateTask methods into one
singular method that works on tasks, projects, and subtasks. Doing this
would require refactoring a lot of other code in the dataManager module,
though, and since I know I am not going to need another category beyond
subtasks I will keep the methods separate. Expanding functionality
of this application later is not a concern in this case.