Refactor toggleShowHide for project elements
Now will allow for a paired element that will first check current
element settings to set itself appropriately so that the two elements
are always opposite of each other with respect to show/hide.

Also removed event.on for reloading projects container on toggleBtn.
This was causing the edit project label to not have intended behavior.